+++ src/main/java/blackjack/model/card/CardNumber.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.card;
 
 import java.util.Arrays;
 import java.util.List;
